Move timestamp to TUF package
